requirements.txt
numpy
numpy-stl
pyrr
PyQt5
Pillow
nbt
trimesh
qdarkstyle
setuptools
PyOpenGl